refactor: move records to mapper namespace
