+++ Connectors/Rhino/Speckle.Connectors.RhinoShared/Mapper/Revit/LayerCategoryMapping.cs
@@ -1,0 +1,12 @@
+namespace Speckle.Connectors.Rhino.Mapper.Revit;
+
+/// <summary>
+/// Represents layers that are all assigned to the same category.
+/// </summary>
+public record LayerCategoryMapping(
+  string CategoryValue,
+  string CategoryLabel,
+  IReadOnlyList<string> LayerIds,
+  IReadOnlyList<string> LayerNames,
+  int LayerCount
+);
