+++ Connectors/Rhino/Speckle.Connectors.RhinoShared/Mapper/Revit/LayerOption.cs
@@ -1,0 +1,6 @@
+namespace Speckle.Connectors.Rhino.Mapper.Revit;
+
+/// <summary>
+/// Represents a layer option for the UI dropdown.
+/// </summary>
+public record LayerOption(string Id, string Name);
